requirements.txt
requests==1.2.3
python-dateutil==2.1